salesforce to AWL ETL support
This script was written to take a CSV exported from Salesforce, scrub it of PII, and reorganize the columns to satisfy ingestion requirements from a legacy system as an interim solution while a formal Salesforce to AWS ingestion process was still under development. It's use began in June 2022 and is anticipated to persist to September 2022, to be retired when the new data pipeline is in place.