Remote Visual Guidance ETL Processor
This script was written to answer a request from the business to measure the use of RVG platform Blitzz by agents as a performance metric. The data did not come prepared to score agents on frequency of use other than by session ID, but as a single conversation might feature several sessions the business requested that we narrow the number of points an agent can get. To do this, this script imposes the logic that for each customer, every agent can score just one incident of RVG use per day. In this way we can realistically measure how proactive the agent is in offering this service without overcounting when technical difficulties or user error necessitate multiple sessions to be used in a single customer interaction.

The script also formats the data so that it can be easily added to an access database that serves its data and calculations on performance to excel-based agent performance reports.